Move file related files to file folder
