+++ packages/interbit/src/file/updateIndexHtml.js
